requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-phonenumber-field==5.0.0
gunicorn==20.0.4
phonenumbers==8.12.11
Pillow==8.0.0
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
